Replace Jackson with Gson
* Create JsonParser to unify logic
* Pros:
- Manifest deserialization reduced from ~590 to ~178ms
- Binary size reduced from 70 to 62MB
* Cons:
- Deserialization of numbers uses Long instead of Integer
Note: raw manifest loading takes 5ms, so that can be ignored
Note: gson still requires reachability-metadata for records

+++ src/main/java/org/mnm/api/HttpHelper.java
@@ -14,7 +14,7 @@ public class HttpHelper {
             throw exception(response);
         }
         Map<String, Object> body = response.body();
-        Integer status = (Integer) body.get("status");
+        Long status = (Long) body.get("status");
         if (status != 0) {
             throw exception(response);
         }
